Fold scores export materialization into the catalog module.
Remove the thin export_materialize module so tree assembly lives next to catalog wiring.
