+++ packages/api/api/analytics/scores/exports.py
@@ -15,13 +15,14 @@ from api.analytics.military_score_inference.inference_stream_scope import Infere
 from api.analytics.military_score_inference.inference_table_stream_registry import (
     controller_for_scope,
 )
-from api.analytics.scores.export_materialize import build_scores_export_materialized_tree
 from api.analytics.scores.export_precedence import (
     ScoresExportResolved,
+    SearchStatus,
     is_persistable_inference_status,
     is_scores_export_authoritatively_persisted,
     is_scores_inference_ensure_satisfied,
     resolve_scores_export,
+    resolve_scores_export_payload,
 )
 from api.analytics.scores.export_schema import EXPORT_VALUE_SCHEMA
 from api.analytics.scores.export_services import ScoresExportContext, resolve_scores_services
@@ -215,6 +216,55 @@ def _ensure_current_turn_scheduler(
         stream_token=inputs.stream_token,
     )
     return True
+
+
+def _export_meta_branch(
+    *,
+    search_status: SearchStatus,
+    host_turn: int,
+    solutions_held: int = 0,
+) -> dict[str, object]:
+    meta: dict[str, object] = {
+        "searchStatus": search_status,
+        "hostTurn": host_turn,
+    }
+    if solutions_held > 0:
+        meta["solutionsHeld"] = solutions_held
+    return meta
+
+
+def _hull_catalog_mask_branch(enabled_hull_ids: frozenset[int] | set[int]) -> dict[str, object]:
+    return {"enabledHullIds": sorted(enabled_hull_ids)}
+
+
+def build_scores_export_materialized_tree(
+    resolved: ScoresExportResolved,
+    scope: ExportScope,
+    *,
+    services: ScoresExportContext,
+    turn: TurnInfo,
+) -> dict[str, Any]:
+    """Materialize the full scores export value tree for one resolved snapshot."""
+    payload = resolve_scores_export_payload(resolved)
+    tree: dict[str, Any] = {
+        "meta": _export_meta_branch(
+            search_status=resolved.decision.search_status,
+            host_turn=scope.turn,
+            solutions_held=payload.solutions_held,
+        ),
+        "solutions": payload.solutions,
+    }
+    if payload.diagnostics is not None:
+        tree["diagnostics"] = payload.diagnostics
+
+    if scope.player_id is not None:
+        resolved_mask = services.resolve_hull_catalog_mask(turn, scope.player_id)
+        if resolved_mask is not None:
+            tree["hullCatalogMask"] = _hull_catalog_mask_branch(
+                resolved_mask.effective_enabled_hull_ids
+            )
+
+    return tree
 
 
 def materialize_scores_export_tree(ctx: AnalyticQueryContext, scope: ExportScope) -> dict[str, Any]:
